+++ src/math/backends/profiler.ts
@@ -1,3 +1,20 @@
+/**
+ * @license
+ * Copyright 2018 Google Inc. All Rights Reserved.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * =============================================================================
+ */
+
 import * as util from '../../util';
 import {NDArray} from '../ndarray';
 import {TypedArray} from '../types';
@@ -25,9 +42,9 @@ export class Profiler {
     const vals = result.dataSync();
     util.checkForNaN(vals, result.dtype, name);
 
-    timer.then(
-        (timeMs: number) =>
-            this.logger.logKernelProfile(kernelName, result, vals, timeMs));
+    timer.then((timeMs: number) => {
+      this.logger.logKernelProfile(kernelName, result, vals, timeMs);
+    });
 
     return result as T;
